Update the congestion control mechanism

+++ server/include/utils.hpp
@@ -15,13 +15,18 @@
 #include <sstream>
 
 class Utils {
-  public:
+public:
     static const int SOCKET_ERROR = -1;
 
-    static DataPacket* createPacket(uint16_t len, uint16_t seqno, std::string& data, int index);
-    static std::vector<DataPacket*> divideFileIntoPackets(std::string fileName);
-    static int sendDataPacket(int socket, struct DataPacket *dataPacket, struct sockaddr_in dest_addr, double loss_probability);
-    static ssize_t send_with_packet_loss(double prob_of_drop, int socket, sockaddr_in &client_address, char const *data, size_t len);
+    static DataPacket *createPacket(int len, int seqno, std::string &data, int index);
+
+    static std::vector<DataPacket *> divideFileIntoPackets(std::string fileName);
+
+    static int
+    sendDataPacket(int socket, struct DataPacket *dataPacket, struct sockaddr_in dest_addr, double loss_probability);
+
+    static ssize_t
+    send_with_packet_loss(double prob_of_drop, int socket, sockaddr_in &client_address, char const *data, size_t len);
 };
 
 #endif /* ifndef UTILS_HEADER */
